refactoring

+++ Hackerspaces/HSJsonParser.swift
@@ -50,7 +50,7 @@ func parseHackerspaceDataModel(json: [String: JSONValue], name apiName: String, 
         let logo = json[SpaceAPIConstants.APIlogo.rawValue]?.asString,
         let websiteURL = json[SpaceAPIConstants.APIurl.rawValue]?.asString,
         let state = json[SpaceAPIConstants.APIstate.rawValue]?.asObject >>- parseStateObject,
-        let location = json[SpaceAPIConstants.APIlocation.rawValue]?.asObject >>- { loc in name >>- {n in parseLocationObject(loc, withName: n) } },
+        let location = json[SpaceAPIConstants.APIlocation.rawValue]?.asObject >>- { loc in name >>- { n in parseLocationObject(loc, withName: n) } },
         let contact = json[SpaceAPIConstants.APIcontact.rawValue]?.asObject >>- parseContactObject,
         let reportChannel = json[SpaceAPIConstants.APIreport.rawValue]?.asArray >>-  parseReportChannel {
 
@@ -91,11 +91,11 @@ private func parseIconObject(_ icon: [String : JSONValue]) -> IconObject? {
 }
 
 func parseLocationObject(_ location: [String : JSONValue], withName name: String) -> SpaceLocation? {
-    let lat = location[JSONKeys.lat]?.asInt >>- CLLocationDegrees.init
-    let lon = location[JSONKeys.lon]?.asInt >>- CLLocationDegrees.init
-    let loc = lat >>- {la in lon >>- { lo in CLLocationCoordinate2D(latitude: la, longitude: lo)}}
+    let lat = location[JSONKeys.lat]?.asFloat >>- CLLocationDegrees.init
+    let lon = location[JSONKeys.lon]?.asFloat >>- CLLocationDegrees.init
+    let loc = lat >>- { la in lon >>- { lo in CLLocationCoordinate2D(latitude: la, longitude: lo) } }
     let addr = location[JSONKeys.address]?.asString
-    return loc >>- {SpaceLocation(name: name, address: addr, location: $0)}
+    return loc >>- { SpaceLocation(name: name, address: addr, location: $0) }
 }
 
 private func parseContactObject(_ contact: [String: JSONValue]) -> ContactObject {
@@ -116,7 +116,7 @@ private func parseContactObject(_ contact: [String: JSONValue]) -> ContactObject
 }
 
 private func parseKeymasters(_ keymasters: [JSONValue]) -> [MemberObject] {
-    let members: [MemberObject?] =  keymasters.map {$0.asObject.map(parseMember)}
+    let members: [MemberObject?] =  keymasters.map { $0.asObject.map(parseMember) }
     return Array(members.joined())
 }
